move CARMA test to examples
